moved KeyIDStr
