+++ services/web/server/src/simcore_service_webserver/projects/projects_api.py
@@ -32,10 +32,11 @@ from models_library.api_schemas_dynamic_scheduler.dynamic_services import (
 )
 from models_library.api_schemas_webserver.projects import ProjectPatch
 from models_library.api_schemas_webserver.projects_nodes import NodePatch
+from models_library.basic_types import KeyIDStr
 from models_library.errors import ErrorDict
 from models_library.products import ProductName
 from models_library.projects import Project, ProjectID, ProjectIDStr
-from models_library.projects_nodes import KeyIDStr, Node, OutputsDict
+from models_library.projects_nodes import Node, OutputsDict
 from models_library.projects_nodes_io import NodeID, NodeIDStr, PortLink
 from models_library.projects_state import (
     Owner,
